refactor: use runRepo.findLatestResultForTest() in testFix route handler
